wip(#1102): extract emit/memcall.rs (rt + emit_memcall*)
